added x11 support
